refactor: Be able to import app
